last commit

+++ Embedded_Final_Project/MCAL/ADC/ADC.h
@@ -3,6 +3,6 @@
 #include "typedefs.h"
 
 void Init_ADC();
-uint16 ADC_Read(uint8 channel);
+uint16 ADC_Read(uint16 channel);
 
 #endif /* ADC_H_ */
